registro y validaciones

+++ model/email.php
@@ -18,25 +18,30 @@
     }
     function enviarCodigo($email){
         global $db;
-        $codigo = generarCodigo();
-        $sql = "INSERT INTO verificaciones VALUES ('$email','$codigo')
-                ON DUPLICATE KEY UPDATE code='$codigo'";
-        $db->query($sql);
-        if($db->error=="")
-            sendMail(
-                $email,
-                "Codigo de verificación",
-                "<h5>Su código de verificación es:</h5>
-                <h2 style=\"color:#007bff\">$codigo</h2>
-                <h6>Si no solicito esto, descarte este mensaje</h6>"
-            );
+        if (validemail($email)){
+            $codigo = generarCodigo();
+            $sql = "INSERT INTO verificaciones VALUES ('$email','$codigo')
+                    ON DUPLICATE KEY UPDATE code='$codigo'";
+            $db->query($sql);
+            if($db->error=="")
+                sendMail(
+                    $email,
+                    "Codigo de verificación",
+                    "<h5>Su código de verificación es:</h5>
+                    <h2 style=\"color:#007bff\">$codigo</h2>
+                    <h6>Si no solicito esto, descarte este mensaje</h6>"
+                );
+        }
     }
     function verificarCodigo($codigo, $email){
         global $db;
-        $sql = "SELECT code FROM verificaciones WHERE email='$email'";
-        if($db->query($sql)->fetch_object()->code == $codigo){
-            $db->query("DELETE FROM verificaciones WHERE email='$email'");
-            return 'true';
-        } return 'false';
+        if(validcode($codigo) & validemail($email)){
+            $sql = "SELECT code FROM verificaciones WHERE email='$email'";
+            if($db->query($sql)->fetch_object()->code == $codigo){
+                $db->query("DELETE FROM verificaciones WHERE email='$email'");
+                return 'true';
+            }    
+        }
+        return 'false';
     }
 ?>
